Heart Print Data
Heartprint is a large biometric database of multisession ECG-signal comprising 1539 records of 15 seconds length collected from 199 subjects. The signals were collected in multiple sessions over ten years starting from 2012 in resting and reading conditions of the subjects and organized in a multisession database. The dataset also covers several demographic classes such as genders, ethnicities, and age groups. Heartprint dataset could be a valuable resource for the development and evaluation of biometric recognition algorithms.